Update Migration Product

+++ database/migrations/2020_05_22_234414_create_products_table.php
@@ -23,7 +23,7 @@ class CreateProductsTable extends Migration
             $table->integer('stock');
 
             //FK
-            $table->unsignedBigInteger('category_id')->unsigned()->nullable();
+            $table->integer('category_id')->unsigned()->nullable();
             $table->foreign('category_id')->references('id')->on('categories');
